requirements.txt
# local package
-e .

# external requirements
click
coverage
flake8
python-dotenv>=0.5.1
torch
torchmetrics
torchvision
numpy
mlflow
